refactor: update package - rename package directory - update imports - change dataset and schema directory
