--- bumpversion.cfg
@@ -1,8 +1,0 @@
-[bumpversion]
-current_version = 0.1.0
-commit = True
-tag = True
-parse = (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)
-serialize = {major}.{minor}.{patch}
-
-[bumpversion:file:pyproject.toml]
