+++ pyproject.toml
@@ -1,9 +1,11 @@
 [build-system]
 requires = ["setuptools>=61.0"]
 build-backend = "setuptools.build_meta"
+# requires = ["hatchling"]
+# build-backend = "hatchling.build"
 
 [project]
-name = "biochattermetta"
+name = "biochatter_metta"
 version = "0.0.1"
 authors = [
   { name="Tollan", email="tollan.ititaw@singularitynet.io" }
