+++ src/biochatter_metta/biocypher_config/bioatomspace/gaf/edges.metta
