+++ src/biochatter_metta/biocypher_config/bioatomspace/gencode/edges.metta
