+++ src/biochatter_metta/biocypher_config/bioatomspace/gtex/eqtl/edges.metta
