+++ src/biochatter_metta/biocypher_config/bioatomspace/ontology/edges.metta
