+++ src/biochatter_metta/biocypher_config/bioatomspace/uniprot/edges.metta
