+++ src/biochatter_metta/biocypher_config/schema_config.yaml
