+++ src/biochatter_metta/biocypher_config/schema_mappings.json
@@ -1,0 +1,3 @@
+{
+    "gene": "./bioatomspace/../..metta"
+}
